+++ codebase/libs/dsserver/src/DsServer/DsProcessServer.cc
@@ -455,11 +455,11 @@ int DsProcessServer::handleServerCommand(Socket * socket,
 
   // get comm timeout
   
-  int commTimeoutMsecs = DS_DEFAULT_COMM_TIMEOUT_MSECS;
+  ssize_t commTimeoutMsecs = DS_DEFAULT_COMM_TIMEOUT_MSECS;
   char *DS_COMM_TIMEOUT_MSECS = getenv("DS_COMM_TIMEOUT_MSECS");
   if (DS_COMM_TIMEOUT_MSECS != NULL) {
-    int timeout;
-    if (sscanf(DS_COMM_TIMEOUT_MSECS, "%d", &timeout) == 1) {
+    ssize_t timeout;
+    if (sscanf(DS_COMM_TIMEOUT_MSECS, "%ld", &timeout) == 1) {
       commTimeoutMsecs = timeout;
     }
   }
@@ -938,11 +938,11 @@ void * DsProcessServer::__serveClient(void * svrsockstruct)
 
   // get comm timeout
   
-  int commTimeoutMsecs = DS_DEFAULT_COMM_TIMEOUT_MSECS;
+  ssize_t commTimeoutMsecs = DS_DEFAULT_COMM_TIMEOUT_MSECS;
   char *DS_COMM_TIMEOUT_MSECS = getenv("DS_COMM_TIMEOUT_MSECS");
   if (DS_COMM_TIMEOUT_MSECS != NULL) {
-    int timeout;
-    if (sscanf(DS_COMM_TIMEOUT_MSECS, "%d", &timeout) == 1) {
+    ssize_t timeout;
+    if (sscanf(DS_COMM_TIMEOUT_MSECS, "%ld", &timeout) == 1) {
       commTimeoutMsecs = timeout;
     }
   }
